1219 fix bugs

+++ window_scene/test/unittest/session_stub_property_test.cpp
@@ -71,7 +71,7 @@ HWTEST_F(SessionStubPropertyTest, HandleUpdatePropertyByAction01, Function | Sma
     MessageOption option{MessageOption::TF_SYNC};
     data.WriteInterfaceToken(u"OHOS.ISession");
     data.WriteUint64(static_cast<uint64_t>(WSPropertyChangeAction::ACTION_UPDATE_MAIN_WINDOW_TOPMOST));
-    uint64_t code = static_cast<uint64_t>(SessionInterfaceCode::TRANS_ID_UPDATE_SESSION_PROPERTY);
+    uint32_t code = static_cast<uint32_t>(SessionInterfaceCode::TRANS_ID_UPDATE_SESSION_PROPERTY);
 
     int ret = session_->OnRemoteRequest(code, data, reply, option);
     ASSERT_EQ(ERR_NONE, ret);
